Get basic resume indicator functionality working
I made the post("/register/{ticket}) also return a copy of the uploaded user data afterwards so I can make the frontend show the indicator right away. Not sure why I decided to not implement that while working on it.

Also, I had to make a new typedict for pylance and fastapi to stop complaining about the post(...) returning the user.

Next: the tests are setup for a table without the resume_filename column, so we'll probably have to change those (and someone who has access needs to alter the table on the remote like I did on mine). I'm also going to make the front end show the indicator immediately after uploading, rather than the next time they go to update their info.