Resume indicator (#66)
* Get basic resume indicator functionality working

I made the post("/register/{ticket}) also return a copy of the uploaded user data afterwards so I can make the frontend show the indicator right away. Not sure why I decided to not implement that while working on it.

Also, I had to make a new typedict for pylance and fastapi to stop complaining about the post(...) returning the user.

Next: the tests are setup for a table without the resume_filename column, so we'll probably have to change those (and someone who has access needs to alter the table on the remote like I did on mine). I'm also going to make the front end show the indicator immediately after uploading, rather than the next time they go to update their info.

* Some Bug Fixes

-added empty resume_filename fields to test models
-removed the type annotation on post response in dashboard (which I guess I added in the first place while trying to match the rest of the functions in dashboard) as that was causing a bunch of FastAPI exceptions.

* Clean up dashboard.py

refactored some of the code I added.

* register.py clean up

Forgot a temporary fix I had in there. Removed it

* Bug Fix: POST returning user with extra field, rather than user and a field

* Return post response to what it was

I had the post send user back to the front end after registering them and uploading everything to update the indicator right away, but the front end just goes to a submit screen so that doesn't matter. Now it's back to how it was.

* post response clean up

final few changes i forgot to include in previous commits

* PR review changes

-Followed the changes requested. Decided to keep the backend returning {user, resume_uploaded_at} separately, as I'm going to be working on the resumes tables and resume review next, so the only resume variable included in user will likely end up being some kind of resume_id linking the 2 tables. Changing what I have to add both as variables in user just to later change them back felt more unnecessary to me than just making it work with them separate(although both ways of handling that would get likely overridden once that branch merges so I guess it doesn't really matter).
- Also changed sqlite.py to reflect my db changes as well. And I'll make sure to note down the exact SQL commands I run on mine in the future!

* Matching frontend naming better

changed resume_uploaded_at to match the front end's resumeUploadedAt state/prop